dmg2img: add LZFSE support, requires HAVE_LZFSE option
DMG files with LZFSE-compressed blocks trigger the following errors:

    Unsupported or corrupted block found: -2147483641

Add support for it. The 4x output size refers to the comment in
https://github.com/lzfse/lzfse/blob/lzfse-1.0/src/lzfse_main.c#L241

The lzfse package is not commonly included with Linux distributions,
therefore it is not enabled by default. Once the library is available,
build with: `make dmg2img HAVE_LZFSE=1`

A sample file can be found at
https://github.com/pdso/DeveloperDiskImage/raw/master/14.3/DeveloperDiskImage.dmg